third_party/blink/web_tests/fast/js/mozilla/strict/11.13.2-expected.txt

PASS Function("'use strict'; arguments+=1") threw exception of type SyntaxError.
PASS !!Function("arguments+=1") is true
PASS true === true
PASS Function("'use strict'; eval+=1") threw exception of type SyntaxError.
PASS !!Function("eval+=1") is true
PASS true === true
PASS Function("'use strict'; (arguments)+=1") threw exception of type SyntaxError.
PASS !!Function("(arguments)+=1") is true
PASS true === true
PASS Function("'use strict'; (eval)+=1") threw exception of type SyntaxError.
PASS !!Function("(eval)+=1") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
